% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_template.R
\name{build_template}
\alias{build_template}
\title{FUNCTION_TITLE}
\usage{
build_template(file, job_type = "test", transfer = "YES",
  transfer_time = "ON_EXIT", args = "$(Process)", tag = "job",
  init_dir = "jobs", input_files = NULL, output_files = NULL,
  jobs = 1, template_file = NULL)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{job_type}{PARAM_DESCRIPTION, Default: 'test'}

\item{transfer}{PARAM_DESCRIPTION, Default: 'YES'}

\item{transfer_time}{PARAM_DESCRIPTION, Default: 'ON_EXIT'}

\item{args}{PARAM_DESCRIPTION, Default: '$(Process)'}

\item{tag}{PARAM_DESCRIPTION, Default: 'job'}

\item{init_dir}{PARAM_DESCRIPTION, Default: 'jobs'}

\item{input_files}{PARAM_DESCRIPTION, Default: NULL}

\item{output_files}{PARAM_DESCRIPTION, Default: NULL}

\item{jobs}{PARAM_DESCRIPTION, Default: 1}

\item{template_file}{character, path to save the template file, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[whisker]{whisker.render}}
}
