% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.qscore.stability.R
\name{calc.qscore.stability}
\alias{calc.qscore.stability}
\title{Classifies each node as either stable or variable between two conditions}
\usage{
calc.qscore.stability(x, y, type = c("red", "blue"), by = c("column", "row",
  "both"))
}
\arguments{
\item{x}{first output of \code{\link{condor.qscore}}}

\item{y}{second output of \code{\link{condor.qscore}}}

\item{type}{character indicating whether overlaps should be determined for blue or red nodes}

\item{by}{character indicating whether to determine stability by 'row', 'column', or 'both'.
Default is "column"}
}
\value{
data.frame with core scores of both conditions and stability
}
\description{
Classifies each node as either stable (remaining in the most shared
community between conditions) or variable (switching communities)
}

