% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_local}
\alias{cleanup_local}
\title{cleanup files on the local machine}
\usage{
cleanup_local(dir, tag)
}
\arguments{
\item{dir}{character, path where the jobs were returned on the local machine}

\item{tag}{character, tag name given to the jobs.}
}
\description{
post processing function to cleanup files used
for the jobs on the local machine.
}
\examples{
\dontrun{
cleanup_local('path','tag')
}
}
