% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_remote}
\alias{cleanup_remote}
\title{cleanup files on the cluster}
\usage{
cleanup_remote(session, dir = "jobs/run")
}
\arguments{
\item{session}{session connection object}

\item{dir}{character, path that jobs were invoked from on the master, Default: 'jobs/run'}
}
\value{
ssh result from the call
}
\description{
post processing function to cleanup files used
for the jobs on the master.
}
\examples{
\dontrun{
cleanup_remote(session,'jobs/run')
}
}
\seealso{
\code{\link[ssh]{ssh_exec}}
}
