% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.compare.qscores.R
\name{condor.compare.qscores}
\alias{condor.compare.qscores}
\title{Plots boxplots of core score distributions by condition and by stability}
\usage{
condor.compare.qscores(qscore, lab.x = "Condition x", lab.y = "Condition y",
  scale.log = FALSE, nsamp = 10000)
}
\arguments{
\item{qscore}{Output of \code{\link{calc.qscore.stability}}}

\item{lab.x}{Label of first condition}

\item{lab.y}{Label of second condition}

\item{scale.log}{TRUE/FALSE - If TRUE, plot log core scores}

\item{nsamp}{Number of permutation tests to run, passed to \code{\link{condor.core.enrich}}}
}
\value{
ggplot object
}
\description{
This function tests whether core scores of stable nodes are higher than those of variable nodes
}

