% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.plot.community.overlap.R
\name{condor.plot.community.overlap}
\alias{condor.plot.community.overlap}
\title{Plots overlap between communities for two condor objects with identical nodes}
\usage{
condor.plot.community.overlap(x, y, main = "", xlab = "Condition 1",
  ylab = "Condition 2", reorder = TRUE)
}
\arguments{
\item{x}{First condor.object membership data.frame}

\item{y}{Second condor.object membership data.frame}

\item{main}{Title of plot}

\item{reorder}{If FALSE, do not reorder rows and columns}
}
\value{
matrix containing overlap between each pair of communities
}
\description{
Plots overlap between communities for two condor objects with identical nodes
}

