% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{condor_permissions}
\alias{condor_permissions}
\title{set permissions for files on the cluster}
\usage{
condor_permissions(session, file, permission = "a+x")
}
\arguments{
\item{session}{session connection object}

\item{file}{character, filename to set permissions}

\item{permission}{character, permissions to set, Default: 'a+x'}
}
\value{
ssh result from the call
}
\description{
wrapper that sets permissions for files on the cluster
}
\examples{
\dontrun{
condor_permissions(session,'file.r','a+x')
}
}
\seealso{
\code{\link[ssh]{ssh_exec}}
}
