% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{condor_rm}
\alias{condor_rm}
\title{condor_rm}
\usage{
condor_rm(session, jobs)
}
\arguments{
\item{session}{session connection object}

\item{jobs}{character/numeric, jobs to kill}
}
\value{
result of call
}
\description{
invokes condor_rm on the cluster
}
\examples{
\dontrun{
condor_rm(session,5000) # set of jobs
condor_rm(session,5000.1) # specific job
}
}
\seealso{
\code{\link[ssh]{ssh_exec}}
}
