% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directories on the cluster}
\usage{
create_dirs(session, file)
}
\arguments{
\item{session}{session connection object}

\item{file}{character, path to condor file on local machine}
}
\value{
ssh result from the call
}
\description{
Create relevant directories for job outputs based on the
condor.file fields.
}
\examples{
\dontrun{
create_dirs(session,'run.condor')
}
}
\seealso{
\code{\link[ssh]{ssh_exec}}
}
