% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_check_jobs.R
\name{make_check_jobs}
\alias{make_check_jobs}
\title{Build bash file to track jobs on Condor}
\usage{
make_check_jobs(session, sleeptime = "10s",
  email = Sys.getenv("UCONN_EMAIL"),
  subject = "auto email from stats cluster",
  body = "All jobs have been completed.  Cheers!")
}
\arguments{
\item{session}{session object}

\item{sleeptime}{character, time interval to check if jobs are still running,
Default: '10s'}

\item{email}{character, email address to send message on end of jobs,
Default: Sys.getenv("UCONN_EMAIL")}

\item{subject}{character, subject line of the email, Default: 'auto email from stats cluster'}

\item{body}{character, body contents of the email,
Default: 'All jobs have been completed.  Cheers!'}
}
\description{
Progrmatically build file to track Condor runs on the cluster.
}
\examples{
\dontrun{
make_check_jobs(session)
}
}
\seealso{
\code{\link[whisker]{whisker.render}}
\code{\link[ssh]{scp}}
}
