% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull}
\alias{pull}
\title{pull files from cluster to local machine}
\usage{
pull(session, from, to)
}
\arguments{
\item{session}{session connection object}

\item{from}{character, path on cluster}

\item{to}{character, path on local machine}
}
\value{
ssh result from the call
}
\description{
wrapper of \link[ssh:scp_download]{scp_download} to pull files from
cluster to the local machine
}
\examples{
\dontrun{
pull(session, from = 'cluster_file.R', to = 'local_file.R')
}
}
\seealso{
\code{\link[purrr]{map2}}
\code{\link[ssh]{scp}}
}
