% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_errs.R
\name{read_errs}
\alias{read_errs}
\title{Summary of err files}
\usage{
read_errs(err_path = "output/err")
}
\arguments{
\item{err_path}{character, path to err subdirectory, Default: 'output/err'}
}
\value{
character
}
\description{
Post processing summary containing all the unique errors and warnings
found in the err subdirectory.
}
