% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssh_fn}
\alias{ssh_fn}
\title{run terminal command on cluster}
\usage{
ssh_fn(session, .f = "whoami", intern = FALSE, convert_intern = TRUE,
  ...)
}
\arguments{
\item{session}{session connection object}

\item{.f}{command to run, Default: 'whoami'}

\item{intern}{boolean, switch between \link[ssh:ssh_exec_internal]{ssh_exec_internal}
and \link[ssh:ssh_exec_wait]{ssh_exec_wait}}

\item{convert_intern}{boolean, convert raw output from
\link[ssh:ssh_exec_internal]{ssh_exec_internal} into human readable objects, Default: TRUE}

\item{\dots}{parameter to pass to ssh function}
}
\value{
ssh result from the command
}
\description{
wrapper to run command on the terminal of the cluster
}
\examples{
\dontrun{
ssh_fn(session,.f='ls ~')
ssh_fn(session,.f='ls ~',intern = TRUE)
}
}
\seealso{
\code{\link[ssh]{ssh_exec}}
}
