% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssh_print}
\alias{ssh_print}
\title{print for ssh_exec_internal call}
\usage{
ssh_print(x)
}
\arguments{
\item{x}{ssh_exec_internal output}
}
\value{
character
}
\description{
converts raw output into characters
}
