% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc.qscore.stability.R
\name{calc.qscore.stability}
\alias{calc.qscore.stability}
\title{Classifies each node as either stable or variable between two conditions}
\usage{
calc.qscore.stability(res, by = c("both", "row", "column"))
}
\arguments{
\item{res}{Output of \code{\link{condor.plot.community.overlap}}}

\item{by}{chaacter string indicating whether to determine stability by 'row', 'column', or 'both'}
}
\value{
A data frame with modularity of both conditions and stability
}
\description{
Classifies each node as either stable or variable based on the output of \code{\link{condor.plot.community.overlap}}
}

