% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.compare.qscores.R
\name{condor.compare.qscores}
\alias{condor.compare.qscores}
\title{Plots boxplots of core score distributions by condition and by stability}
\usage{
condor.compare.qscores(x, y, z, lab.x = "Condition 1",
  lab.y = "Condition 2", nsamp = 1000)
}
\arguments{
\item{x}{First condor.object qscores data.frame}

\item{y}{Second condor.object qscores data.frame}

\item{z}{Output of \code{\link{condor.plot.community.overlap}}}

\item{lab.x}{Label of first condition}

\item{lab.y}{Label of second condition}

\item{nsamp}{Number of permutation tests to run, passed to \code{\link{condor.core.enrich}}}
}
\value{
ggplot object
}
\description{
This function tests whether core scores of stable nodes are higher than those of variable nodes
}

