% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.compare.qscores.R
\name{condor.compare.qscores}
\alias{condor.compare.qscores}
\title{Plots boxplots of core score distributions by condition and by stability for both red and blue nodes}
\usage{
condor.compare.qscores(x, y, by = c("column", "row", "both"),
  label.x = "Condition x", label.y = "Condition y", red.name = "red",
  blue.name = "blue", scale.log = FALSE, nsamp = 10000)
}
\arguments{
\item{x}{first output of \code{\link{condor.qscore}}}

\item{y}{second output of \code{\link{condor.qscore}}}

\item{by}{character indicating whether to determine stability by 'row', 'column', or 'both'}

\item{label.x}{Label of first condition}

\item{label.y}{Label of second condition}

\item{scale.log}{TRUE/FALSE - If TRUE, plot log core scores}

\item{nsamp}{Number of permutation tests to run, passed to \code{\link{condor.core.enrich}}}
}
\value{
ggplot object
}
\description{
This function tests whether core scores of stable nodes are higher than those of variable nodes
}

