% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor.compare.qscores.R
\name{condor.compare.qscores}
\alias{condor.compare.qscores}
\title{Plots boxplots of core score distributions by condition and by stability for red and/or blue nodes}
\usage{
condor.compare.qscores(x, y, by = c("column", "row", "both"),
  label.x = "Condition x", label.y = "Condition y", red.name = "red",
  blue.name = "blue", scale.log = FALSE, nsamp = 10000,
  plot.type = c("violin", "points"), node.type = c("both", "red", "blue"),
  notch = TRUE)
}
\arguments{
\item{x}{first output of \code{\link{condor.qscore}}}

\item{y}{second output of \code{\link{condor.qscore}}}

\item{by}{character string indicating whether to determine stability by 'row', 'column', or 'both'}

\item{label.x}{Label of first condition}

\item{label.y}{Label of second condition}

\item{red.name}{Name for red nodes}

\item{blue.name}{Name for blue nodes}

\item{scale.log}{TRUE/FALSE - If TRUE, plot log core scores}

\item{nsamp}{Number of permutation tests to run, passed to \code{\link{condor.core.enrich}}}

\item{plot.type}{character string indicating whether to plot violin plots or points}

\item{node.type}{character string indicating whether to plot core score comparisons for red nodes, blue nodes, or both}

\item{notch}{TRUE/FALSE - Add notches to boxplots}
}
\value{
ggplot object
}
\description{
This function tests whether core scores of stable nodes are higher than those of variable nodes
}

