% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.plot.community.overlap.R
\name{condor.plot.community.overlap}
\alias{condor.plot.community.overlap}
\title{Plots overlap between communities for two condor objects with identical nodes}
\usage{
condor.plot.community.overlap(x, y, type = c("red", "blue"),
  scale = c("none", "row", "column"))
}
\arguments{
\item{x}{first output of \code{\link{condor.cluster}},
\code{\link{condor.modularity.max}}, or \code{\link{condor.qscore}}}

\item{y}{second output of \code{\link{condor.qscore}},
\code{\link{condor.modularity.max}}, or \code{\link{condor.qscore}}}

\item{type}{character indicating whether overlaps should be determined for blue or red nodes}

\item{scale}{character indicating if the values should be scaled proportionally in either the row direction or the column direction, or none. The default is \code{"none"}.}
}
\value{
produces a ggplot object
}
\description{
Plots overlap between communities for two condor objects with identical nodes
}

