% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.plot.community.overlap.R
\name{condor.plot.community.overlap}
\alias{condor.plot.community.overlap}
\title{Plots overlap between communities for two condor objects with identical nodes}
\usage{
condor.plot.community.overlap(x, y, xlab = "Condition y",
  ylab = "Condition x", scale.condition = c("none", "row", "column"), ...)
}
\arguments{
\item{x}{First condor.object membership or qscores data.frame}

\item{y}{Second condor.object membership or qscores data.frame}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{other arguments passed to heatmap.2}

\item{scale}{character indicating if the values should be scaled in either the row direction or the column direction, or none. The default is \code{"none"}.}
}
\value{
matrix containing overlap between each pair of communities
}
\description{
Plots overlap between communities for two condor objects with identical nodes
}

