% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condor.plot.heatmap.R
\name{condor.plot.heatmap}
\alias{condor.plot.heatmap}
\title{Plot weighted adjacency matrix with links grouped by community}
\usage{
condor.plot.heatmap(condor.object, community.ticks = FALSE,
  add.color = FALSE, colors = NULL, log = FALSE, xlab = "blues",
  ylab = "reds", main = "", tick.min.prop = 0.05, xlabels = NULL,
  ylabels = NULL, legend.position = "right")
}
\arguments{
\item{condor.object}{output of either \code{\link{condor.cluster}} or
\code{\link{condor.modularity.max}}}

\item{community.ticks}{if TRUE, plot community numbers in x and y axes rather than column and row names}

\item{add.color}{if TRUE, fill in communities with solid colors}

\item{colors}{character vector of colors of length equal to the number of communities
in the network}

\item{log}{if TRUE, transform the color gradient to log-scale.}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{main}{a title for the plot}

\item{tick.min.prop}{minimum proportion of genes/TFs a community must contain to have a visible tick mark, used to prevent overlap}

\item{xlabels}{optional character vector of length equal to the number of communities to manually mark the x-axis community ticks}

\item{ylabels}{optional character vector of length equal to the number of communities to manually mark the y-axis community ticks}

\item{legend.position}{side the legend is placed}
}
\value{
produces a ggplot object if \code{color} is FALSE, else no return value
}
\description{
This function will generate the network link 'heatmap' for a weighted network
}
\examples{
data(small1976)
condor.object <- create.condor.object(small1976)
condor.object <- condor.cluster(condor.object, project=FALSE)
condor.plot.heatmap(condor.object)
}

