% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_errs.R
\name{read_errs}
\alias{read_errs}
\title{FUNCTION_TITLE}
\usage{
read_errs(err_path = "output/err")
}
\arguments{
\item{err_path}{PARAM_DESCRIPTION, Default: 'output/err'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
