% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_print.R
\name{ssh_print}
\alias{ssh_print}
\title{FUNCTION_TITLE}
\usage{
ssh_print(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
